+++ src/main/java/roomescape/model/ReservationDto.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.model;
 
 public record ReservationDto(
         Long id,
