+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,7 +1,9 @@
-package roomescape;
+package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.repository.ReservationTimeRepository;
+import roomescape.model.ReservationTime;
 
 @Service
 public class ReservationTimeService {
